server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
apscheduler==3.10.4
cryptography==43.0.1
pyyaml==6.0.2
httpx==0.27.2
pytz==2024.1
pydantic==2.9.2
pydantic-settings==2.5.2
